Figures

Description

Defines all the figures that can be placed in the simulation.

Documentation

figureForType

Arguments

:: FigureType
The FigureType

-> Figure

Returns the figure for a given FigureType

nextFig

Arguments

:: FigureType
The type to get the successor of

-> FigureType
The successor

Returns the FigureType that is succeeding the given type

This is the order:

None

Block

BeeHive

Loaf

Boat

Tub

Blinker

Toad

Beacon

Pulsar

Pentadecathlon

Glider

LightWeight

MiddleWeight

HeavyWeight

Clock

DoubleU

RPentomino

BlockRow

LittleSwimmer

Swimmer

WichStretcher

GliderGun1

GliderGun2

toFigure :: FigureDescription -> Figure

Turns a FigureDescription into a Figure

block :: Figure

The figure for a still block

beeHive :: Figure

The figure for a still bee hive

loaf :: Figure

The figure for a still loaf

boat :: Figure

The figure for a still boat

tub :: Figure

The figure for a still tub

blinker :: Figure

The figure for an oscillating blinker

toad :: Figure

The figure for an oscillating toad

beacon :: Figure

The figure for an oscillating beacon

pulsar :: Figure

The figure for an oscillating pulsar

pentadecathlon :: Figure

The figure for an oscillating penthadecathlon

glider :: Figure

The figure for a glider

lightWeight :: Figure

The figure for a light weight spaceship

middleWeight :: Figure

The figure for a middle weight spaceship

heavyWeight :: Figure

The figure for a heavy weight spaceship

clock :: Figure

The figure for a clock

doubleU :: Figure

The figure for a flipped U on top of a U

rPentomino :: Figure

blockRow :: Figure

The figure for a row of blocks

littleSwimmer :: Figure

The figure for a little swimmer

swimmer :: Figure

The figure for a swimmer

wichStretcher :: Figure

The figure for a wich stretcher

gliderGun1Left :: Figure

The figure of the left generator of glider gun 1

gliderGun1Right :: Figure

The figure of the right generator of glider gun 1

gliderGun2Top :: Figure

The figure of the top generator of glider gun 2

gliderGun2Bottom :: Figure

The figure of the bottom generator of glider gun 2