GameOfLife

Description

Contains everything related to Conway's Game of Life.

Documentation

backgroundColor :: Color

The color of the background

cellColor :: Color

The color of the cells

gridColor :: Color

The color of the grid

genRandomCells

Arguments

:: StdGen
The random generator

-> Grid
The resulting random grid

Generates a random grid of cells

randomGame

Arguments

:: StdGen
The random generator

-> GameData
The new game state

Create a new game with a random field

blankGame

Arguments

:: StdGen
The random generator

-> GameData
The new game state

Create a new game with a blank field

newGameData

Arguments

:: StdGen
The random generator

-> Grid
The starting grid

-> Color
The cell color

-> GameData
The resultung game state

Creates a new game state.

windowSize :: (Int, Int)

The initial window size

cellSize :: Int

The size in pixels of a single Cell

cellCount :: (Int, Int)

The cell count in x and y direction

drawCell

Arguments

:: Cell
The cell to draw | The position of the cell, in cell-coordinates rather than pixels

-> Position

-> Color
The cell color

-> Picture
A function to draw the wiregrid

-> Picture
The resulting picture

Draws a single Cell

translatePosition

Arguments

:: Position
The cell position in cell-coordinate

-> Picture
The old picture

-> Picture
The resulting picture, after the translation

Translates a Picture using a cells Position

drawWire

Arguments

:: Color
The wire color

-> Bool
Whether to draw the wire or not

-> Picture
The resulting picture

Draws the wire background.

drawGame

Arguments

:: GameData
The game state

-> Picture
The resulting picture

Draws the given game state.

drawGrid

Arguments

:: Int
The current cell index in the grid

-> Grid
The Grid to draw

-> Color
The cell color

-> Color
The grid color

-> Bool
Whether to show the wiregrid

-> Picture
The resulting Picture

Draws a Grid

insertGliderGun1

Arguments

:: Grid
The grid to insert the glider gun into

-> Position
The position to insert at

-> Grid
The resulting grid, after the insertion

Inserts the glider gun 2 into a given Grid at a Position

insertGliderGun2

Arguments

:: Grid
The grid to insert the glider gun into

-> Position
The position to insert at

-> Grid
The resulting grid, after the insertion

Inserts the glider gun 1 into a given Grid at a Position

getCellAtPos

Arguments

:: Grid
The Grid the cell is in

-> Position
The Position the cell is at

-> Cell
The resulting cell

Returns the Cell at a given Position in a given Grid.

simulateCell

Arguments

:: Cell
The cell to simulate

-> Int
The number of alive neighbors

-> Cell
The resulting cell, after the simulation

Simulates a Cell with Conway's rules.

These are the rules:

An Alive cell lives with 2 or 3 alive neighbors

A Dead cell lives with exactly 3 alive neighbors

A cell with less than 2 or more than 3 neighbors dies

simulateGrid

Arguments

:: Grid
The grid to simulate

-> Int
The index of the current cell

-> Grid
The resuling grid, after the simulation

Simulates a Grid with Conway's rules.

For exact rules see simulateCell

countCell

Arguments

:: Cell
The cell to count | The resulting number. 1 if the cell is Alive, 0 otherwise.

-> Int

Counts a single Cell.

countAliveNeighbors

Arguments

:: Position
The position the cell is at

-> Grid
The grid the cell is in

-> Int
The amount of alive neighbors

Returns the number of neighboring cells that are Alive.

insertFigure

Arguments

:: Grid
The grid to insert the figure into

-> Int
The index we are at

-> Position
The position to insert the figure

-> Figure
The figure to insert

-> Grid
The resulting grid, after the insertion

Inserts a Figure at a given Grid and position

insertFigurePart

Arguments

:: Grid
The grid to insert the part into

-> [Cell]
The cells of that part that were already inserted

-> Int
The index in the grid to insert this part

-> FigurePart
The part to insert | The result: (rest of the grid after inserting, the inserted cells, the number of inserted cells)

-> (Grid, [Cell], Int)

Inserts a FigurePart at a given Grid and index

nextColor

Arguments

:: Color
The previous color

-> Color
The next color

Returns the subsequent color to the specified color.

This is used to change the color of the cells.

The order is the following:

white

red

green

blue

yellow

cyan

magenta

rose

violet

orange

handleKeys

Arguments

:: Event
The key event that has occured

-> GameData
The old game state

-> GameData
The new game state, after the key press was handled

Updates the game state depending on pressed keys. The following are interactable keys:

Space -> Pause and Unpause the game

Right Arrow -> Step one iteration forward

F1 -> Clear grid

F2 -> Create a random grid

F12 -> Change cell color (see nextColor)

R -> Reset grid

G -> Show / Hide grid

N -> Select next figure to place

LeftMouseButton -> Place selected figure at cursor

pixelPosToCellPos

Arguments

:: (Float, Float)
The pixel coordinates

-> Position
The resulting cell coordinates

Converts pixel coordinates into cell coordinates

The cell coordinates start with (0, 0) at the top left corner but pixel coordinates start at the center of the window

update

Arguments

:: Float
The time in seconds that has passed since the last call to this function

-> GameData
The old game state

-> GameData
The new game state, after stepping one iteration

Steps the game state one iteration forward.